jason 21055999d intro
Path planning is used in different aspects like aviation, robotics, and computer games. It can let an object move to the destination from the starting point by finding a sequence of valid configurations.
In aviation, it is very common to use path planning, pilot needs to do a path plan before each flight in order to make sure the navigation aid is available, it can keep safe and lower the operation cost like fuel. Besides, atc needs to find the best use of the crowded airspace through path planning, it can keep safe and try to lower the probability of delay.