Update README.md
Path planning is used in different aspects like aviation, robotics, and computer games. It can let an object move to the destination from the starting point by finding a sequence of valid configurations.

In aviation, it is very common to use path planning, pilot needs to do a path plan before each flight in order to make sure the navigation aid is available. Besides, atc needs to find the best use of the crowded airspace through path planning, it can keep safe and try to lower the probability of delay.

Moreover, path planning can bring safety to the aviation industry. Flight safety is the biggest concern in the aviation industry. There are many flight accidents are involved in weather conditions in recent years. It is very normal to observe bad weather conditions during the flight. Good path planning can enhance the efficiency of air traffic management, also it can help the aircraft avoid bad weather conditions, prohibited areas. In addition, a backup plan can also be planned in path planning in order to cope with an emergency situations. Path planning can also find the path which used the least energy cost so it can help the airline to reduce the operation cost, safety and reduce the workload of the pilot and air traffic controller.